Added code to save each province as a different file
I have removed the nesting operation in favor of filtering to generate the data for each province and save them in separate files. I also have added a few lines to specify that the new files should be saved in a folder called provinces_files_folder located in the working directory.